init matching thematic page structure
