+++ expo/ppal/app/(tabs)/_layout.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Tabs } from 'expo-router';
-import { Globe, ChefHat, ShoppingCart } from 'lucide-react-native';
+import { ShoppingCart } from 'lucide-react-native';
 
 export default function TabsLayout() {
   return (
     <Tabs screenOptions={{ headerShown: false }}>
       <Tabs.Screen name="pantry" options={{ title: 'Pantry' }} />
-      <Tabs.Screen name="chats" options={{ title: 'Chats' }} />
-      <Tabs.Screen name="explore" options={{ title: 'Explore' }} />
+      <Tabs.Screen name="cookbook" options={{ title: 'Cookbook' }} />
+      <Tabs.Screen name="discover" options={{ title: 'Discover' }} />
     </Tabs>
   );
 }
